@@ -1,6 +1,6 @@
 from construct.lib.py3compat import *
 import re
-import collections
+import sys
 
 
 globalPrintFullStrings = False
@@ -56,21 +56,16 @@ def recursion_lock(retval="<recursion detected>", lock_name="__recursion_lock__"
     return decorator
 
 
-class Container(collections.OrderedDict):
+class Container(dict):
+    # NOTE: be careful when working with these objects. Any method can be shadowed, so instead of doing `self.items()` you should do `dict.items(self)`. Operation that use methods implicitly (such as `x in self` or `self[k]`) will work as usual.
     r"""
-    Generic ordered dictionary that allows both key and attribute access, and preserves key order by insertion. Adding keys is preferred using \*\*entrieskw (requires Python 3.6). Equality does NOT check item order. Also provides regex searching.
-
-    Note that not all parameters can be accessed via attribute access (dot operator). If the name of an item matches a method name of the Container, it can only be accessed via key acces (square brackets). This includes the following names: clear, copy, fromkeys, get, items, keys, move_to_end, pop, popitem, search, search_all, setdefault, update, values.
+    Generic ordered dictionary that allows both key and attribute access, and preserves key order by insertion. Adding keys is preferred using \*\*entrieskw. Equality does NOT check item order. Also provides regex searching.
 
     Example::
 
-        # empty dict
         >>> Container()
-        # list of pairs, not recommended
-        >>> Container([ ("name","anonymous"), ("age",21) ])
-        # This syntax requires Python 3.6
+        >>> Container([("name", "anonymous"), ("age", 21)])
         >>> Container(name="anonymous", age=21)
-        # copies another dict
         >>> Container(dict2)
         >>> Container(container2)
 
@@ -83,54 +78,28 @@ class Container(collections.OrderedDict):
             text = u'utf8 decoded string...' (total 22)
             value = 123
     """
-    __slots__ = ["__recursion_lock__"]
+    __slots__ = ('__dict__', '__recursion_lock__')
 
-    def __getattr__(self, name):
-        try:
-            if name in self.__slots__:
-                return object.__getattribute__(self, name)
-            else:
-                return self[name]
-        except KeyError:
-            raise AttributeError(name)
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.__dict__ = self
 
-    def __setattr__(self, name, value):
-        try:
-            if name in self.__slots__:
-                return object.__setattr__(self, name, value)
-            else:
-                self[name] = value
-        except KeyError:
-            raise AttributeError(name)
+    def copy(self, /):
+        return self.__class__(self)
 
-    def __delattr__(self, name):
-        try:
-            if name in self.__slots__:
-                return object.__delattr__(self, name)
-            else:
-                del self[name]
-        except KeyError:
-            raise AttributeError(name)
+    def __copy__(self, /):
+        return self.__class__.copy(self)
 
-    def update(self, seqordict):
-        """Appends items from another dict/Container or list-of-tuples."""
-        if isinstance(seqordict, dict):
-            seqordict = seqordict.items()
-        for k,v in seqordict:
-            self[k] = v
+    # this is required because otherwise copy.deepcopy() will
+    # copy self and self.__dict__ separately for some reason
+    def __deepcopy__(self, _, /):
+        return self.__class__.copy(self)
 
-    def copy(self):
-        return Container(self)
-
-    __update__ = update
-
-    __copy__ = copy
-
-    def __dir__(self):
+    def __dir__(self, /):
         """For auto completion of attributes based on container values."""
-        return list(self.keys()) + list(self.__class__.__dict__) + dir(super(Container, self))
+        return list(self.__class__.keys(self)) + list(self.__class__.__dict__) + dir(super(Container, self))
 
-    def __eq__(self, other):
+    def __eq__(self, other, /):
         if self is other:
             return True
         if not isinstance(other, dict):
@@ -140,43 +109,36 @@ class Container(collections.OrderedDict):
                 import numpy
                 return numpy.array_equal(v1, v2)
             return v1 == v2
-        for k,v in self.items():
-            if isinstance(k, unicodestringtype) and k.startswith(u"_"):
-                continue
-            if isinstance(k, bytestringtype) and k.startswith(b"_"):
+        for k, v in self.__class__.items(self):
+            if isinstance(k, str) and k.startswith("_"):
                 continue
             if k not in other or not isequal(v, other[k]):
                 return False
-        for k,v in other.items():
-            if isinstance(k, unicodestringtype) and k.startswith(u"_"):
-                continue
-            if isinstance(k, bytestringtype) and k.startswith(b"_"):
+        for k, v in other.__class__.items(other):
+            if isinstance(k, str) and k.startswith("_"):
                 continue
             if k not in self or not isequal(v, self[k]):
                 return False
         return True
 
-    def __ne__(self, other):
-       return not self == other
+    def __ne__(self, other, /):
+        return not self == other
 
     @recursion_lock()
-    def __repr__(self):
+    def __repr__(self, /):
         parts = []
-        for k,v in self.items():
+        for k, v in self.__class__.items(self):
             if isinstance(k, str) and k.startswith("_"):
                 continue
-            if isinstance(v, stringtypes):
-                parts.append(str(k) + "=" + reprstring(v))
-            else:
-                parts.append(str(k) + "=" + repr(v))
+            parts.append(f'{k}={v!r}')
         return "Container(%s)" % ", ".join(parts)
 
     @recursion_lock()
-    def __str__(self):
+    def __str__(self, /):
         indentation = "\n    "
         text = ["Container: "]
         isflags = getattr(self, "_flagsenum", False)
-        for k,v in self.items():
+        for k, v in self.__class__.items(self):
             if isinstance(k, str) and k.startswith("_") and not globalPrintPrivateEntries:
                 continue
             if isflags and not v and not globalPrintFalseFlags:
@@ -188,28 +150,28 @@ class Container(collections.OrderedDict):
                 text.append("(enum) %s %s" % (v, v.intvalue, ))
             elif v.__class__.__name__ in ["HexDisplayedBytes", "HexDumpDisplayedBytes"]:
                 text.append(indentation.join(str(v).split("\n")))
-            elif isinstance(v, bytestringtype):
+            elif isinstance(v, bytes):
                 printingcap = 16
                 if len(v) <= printingcap or globalPrintFullStrings:
-                    text.append("%s (total %d)" % (reprstring(v), len(v)))
+                    text.append("%s (total %d)" % (repr(v), len(v)))
                 else:
-                    text.append("%s... (truncated, total %d)" % (reprstring(v[:printingcap]), len(v)))
-            elif isinstance(v, unicodestringtype):
+                    text.append("%s... (truncated, total %d)" % (repr(v[:printingcap]), len(v)))
+            elif isinstance(v, str):
                 printingcap = 32
                 if len(v) <= printingcap or globalPrintFullStrings:
-                    text.append("%s (total %d)" % (reprstring(v), len(v)))
+                    text.append("%s (total %d)" % (repr(v), len(v)))
                 else:
-                    text.append("%s... (truncated, total %d)" % (reprstring(v[:printingcap]), len(v)))
+                    text.append("%s... (truncated, total %d)" % (repr(v[:printingcap]), len(v)))
             else:
                 text.append(indentation.join(str(v).split("\n")))
         return "".join(text)
 
-    def _search(self, compiled_pattern, search_all):
+    def _search(self, compiled_pattern, search_all, /):
         items = []
-        for key in self.keys():
+        for key, value in self.__class__.items(self):
             try:
-                if isinstance(self[key], (Container,ListContainer)):
-                    ret = self[key]._search(compiled_pattern, search_all)
+                if isinstance(value, (Container, ListContainer)):
+                    ret = value.__class__._search(value, compiled_pattern, search_all)
                     if ret is not None:
                         if search_all:
                             items.extend(ret)
@@ -217,10 +179,10 @@ class Container(collections.OrderedDict):
                             return ret
                 elif compiled_pattern.match(key):
                     if search_all:
-                        items.append(self[key])
+                        items.append(value)
                     else:
-                        return self[key]
-            except:
+                        return value
+            except Exception:
                 pass
         if search_all:
             return items
@@ -232,14 +194,27 @@ class Container(collections.OrderedDict):
         Searches a container (non-recursively) using regex.
         """
         compiled_pattern = re.compile(pattern)
-        return self._search(compiled_pattern, False)
+        return self.__class__._search(self, compiled_pattern, False)
 
     def search_all(self, pattern):
         """
         Searches a container (recursively) using regex.
         """
         compiled_pattern = re.compile(pattern)
-        return self._search(compiled_pattern, True)
+        return self.__class__._search(self, compiled_pattern, True)
+
+    def __getstate__(self, /):
+        """
+        Used by pickle to serialize an instance to a dict.
+        """
+        return dict(self)
+
+    def __setstate__(self, state, /):
+        """
+        Used by pickle to de-serialize from a dict.
+        """
+        self.__class__.clear(self)
+        self.__class__.update(self, state)
 
 
 class ListContainer(list):
@@ -253,8 +228,10 @@ class ListContainer(list):
 
     ::
 
+        >>> obj
+        ListContainer([1, 2, 3])
         >>> print(repr(obj))
-        [1, 2, 3]
+        ListContainer([1, 2, 3])
         >>> print(obj)
         ListContainer
             1
@@ -263,11 +240,11 @@ class ListContainer(list):
     """
 
     @recursion_lock()
-    def __repr__(self):
-        return "ListContainer(%s)" % (list.__repr__(self), )
+    def __repr__(self, /):
+        return "ListContainer(%s)" % (list.__repr__(self),)
 
     @recursion_lock()
-    def __str__(self):
+    def __str__(self, /):
         indentation = "\n    "
         text = ["ListContainer: "]
         for k in self:
@@ -276,12 +253,12 @@ class ListContainer(list):
             text.append(indentation.join(lines))
         return "".join(text)
 
-    def _search(self, compiled_pattern, search_all):
+    def _search(self, compiled_pattern, search_all, /):
         items = []
         for item in self:
             try:
-                ret = item._search(compiled_pattern, search_all)
-            except:
+                ret = item.__class__._search(item, compiled_pattern, search_all)
+            except Exception:
                 continue
             if ret is not None:
                 if search_all:
