@@ -1,3 +1,3 @@
-version = (2,10,69)
-version_string = "2.10.69"
-release_date = "2023.10.22"
+version = (2,10,70)
+version_string = "2.10.70"
+release_date = "2023.11.29"
